+++ src/interfaces/state.constants.ts
@@ -5,5 +5,6 @@ export enum StateConstants {
   FAILED = 'FAILED',
   SKIPPED = 'SKIPPED',
   ROLLING_BACK = 'ROLLING_BACK',
-  ROLLED_BACK = 'ROLLED_BACK'
+  ROLLED_BACK = 'ROLLED_BACK',
+  RETRY = 'RETRY'
 }